add mavenCentral as priority for repos than jcenter
JCenter has been offline often and it is causing pipeline issues when trying to download the poms/jars. 
Since JCenter also serves as a mirror for Maven Central, and any dependencies available on Maven Central are also available on JCenter, it is better to rely on Maven Central since Gradle team is discouraging the use of JCenter since it became read-only.